+++ crates/teloxide-core/src/types/external_reply_info.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 
 use crate::types::{
-    Animation, Audio, Chat, Contact, Dice, Document, Game, Giveaway, GiveawayWinners, Invoice,
-    LinkPreviewOptions, Location, MessageId, MessageOrigin, PaidMediaInfo, PhotoSize, Poll,
-    Sticker, Story, Venue, Video, VideoNote, Voice,
+    Animation, Audio, Chat, Checklist, Contact, Dice, Document, Game, Giveaway, GiveawayWinners,
+    Invoice, LinkPreviewOptions, Location, MessageId, MessageOrigin, PaidMediaInfo, PhotoSize,
+    Poll, Sticker, Story, Venue, Video, VideoNote, Voice,
 };
 
 /// This object contains information about a message that is being replied to,
@@ -52,6 +52,7 @@ pub enum ExternalReplyInfoKind {
     Location(Location),
     Photo(Vec<PhotoSize>),
     Poll(Poll),
+    Checklist(Checklist),
     Sticker(Sticker),
     Story(Story),
     Giveaway(Giveaway),